debug

+++ CS3343_Project_Pilot/src/cs2115Handler/NANDHandler.java
@@ -6,7 +6,7 @@ public class NANDHandler extends LogicHandler {
 
 	@Override
 	public double handleForDoubleReturn(String expression) throws InvalidExpressionException {
-		return new NOTHandler().handleForDoubleReturn(""+new ANDHandler().handleForDoubleReturn(expression));
+		return new NOTHandler().handleForDoubleReturn(""+(int)new ANDHandler().handleForDoubleReturn(expression));
 	}
 
 }
